Added Branch prediction, handling misprediction using BIS, X register
Added masterStall hack - stall decode and fetch stages whenever a JUMP
and BAL  is encountered. Fetch starts when we know the target address of
the unconditional branch i.e. once the JUMP/BAL finsihes EX.

static branch prediction in Decode stage, when BZ/BNZ is in Decode, we
predict based on the sign of the offset and store this into the
instruction to later detect a branch misprediction when BZ/BNZ is in EX.

BIS-
put a branch address into the BIS stack. Whenever a branch instruction
commits, we reomve this from the BIS.
If misprediction, remove all instructions from the processor which have
branchTag >= to the branchTag for the branch mispredicted
topBIS points to the youngest Branch predicted.
Note - JUMP/BAL never go onto BIS as they are never predicted
(unconditional Branches)

Added logic for updating RAT and freeList on misprediction, reInIt them
and then walkback from head of the ROB to the mispredicted branch and
put the required into RAT and allocated list, also, remove all
subsequent instructions in the ROB
Also, remove instruction from IQ which have branchTags >= the
mispredicted branch

LSQ skeleton,
yet to complete,

Xregister for a BAL
X is an architecture register which has the return address for a BAL
Known Bug- NOT writing back X in WB stage, can cause problems.